requirements.txt
django == 3.2.6
django-cors-headers == 3.8.0
djangorestframework == 3.12.4
pymysql == 1.0.2
